Move passthrough tests into bun-test.test.ts
